requirements.txt
langchain
langchain-community
langchain-groq
langgraph
groq
chromadb
sentence-transformers
rank-bm25
easyocr
opencv-python-headless
pillow
faster-whisper
pydub
soundfile
sympy
numpy
scipy
matplotlib
plotly
streamlit
streamlit-extras
python-dotenv
pydantic
pyyaml
networkx
requests
loguru
pandas
tavily-python>=0.7.21
